Extracting StewardsFile::Steward -> StewardConfig
