+++ lib/ladle/stewards_file.rb
@@ -1,28 +1,8 @@
+require 'ladle/steward_config'
+
 module Ladle
   class StewardsFile
     attr_reader :stewards
-
-    class Steward
-      attr_reader :github_username, :include_patterns, :exclude_patterns
-
-      def initialize(github_username:, include_patterns: nil, exclude_patterns: nil)
-        @github_username = github_username
-        @include_patterns = Array.wrap(include_patterns || [])
-        @exclude_patterns = Array.wrap(exclude_patterns || [])
-      end
-
-      def ==(other)
-        @github_username == other.github_username &&
-          @include_patterns == other.include_patterns &&
-          @exclude_patterns == other.exclude_patterns
-      end
-
-      def hash
-        [@github_username, @include_patterns, @exclude_patterns].hash
-      end
-
-      alias eql? ==
-    end
 
     def initialize(stewards = [])
       @stewards = stewards
@@ -42,7 +22,7 @@ module Ladle
           include_patterns = config["include"]
           exclude_patterns = config["exclude"]
         end
-        Steward.new(github_username: github_username, include_patterns: include_patterns, exclude_patterns: exclude_patterns)
+        StewardConfig.new(github_username: github_username, include_patterns: include_patterns, exclude_patterns: exclude_patterns)
       end
 
       new(stewards)
